elmo_hype/requirements.txt
allennlp-models==v2.0.1
